Refactor Util testing classes
