+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -1,11 +1,11 @@
 package seedu.address.logic;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static seedu.address.logic.commands.CommandTestUtil.COURSE_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.GROUP_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.ID_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.COURSE_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.GROUP_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.ID_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.NAME_DESC_AMY;
 import static seedu.address.logic.parser.CliSyntax.CONTACT_COMMAND_WORD;
 import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalContacts.AMY;
