+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -1,123 +1,20 @@
 package seedu.address.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_COURSE_LONG;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_EMAIL_LONG;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_GROUP_LONG;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_ID_LONG;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_NAME_LONG;
-import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_TAG_LONG;
-import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_DEADLINE_LONG;
-import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_LOCATION_LONG;
-import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_NAME_LONG;
-import static seedu.address.logic.parser.todo.TodoCliSyntax.PREFIX_TODO_TAG_LONG;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import seedu.address.commons.core.Operator;
-import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.update.EditContactDescriptor;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
 import seedu.address.model.contact.ContactManager;
-import seedu.address.model.item.predicate.NamePredicate;
-import seedu.address.testutil.EditContactDescriptorBuilder;
 
 /**
  * Contains helper methods for testing commands.
  */
 public class CommandTestUtil {
-
-    // For contact
-    public static final String VALID_ID_AMY = "B0000000001";
-    public static final String VALID_ID_BOB = "B0000000002";
-    public static final String VALID_NAME_AMY = "Amy Bee";
-    public static final String VALID_NAME_BOB = "Bob Choo";
-    public static final String VALID_EMAIL_AMY = "amy@example.com";
-    public static final String VALID_EMAIL_BOB = "bob@example.com";
-    public static final String VALID_ADDRESS_AMY = "Block 312, Amy Street 1";
-    public static final String VALID_TAG_HUSBAND = "husband";
-    public static final String VALID_TAG_FRIEND = "friend";
-    public static final String VALID_COURSE_AMY = "CS50";
-    public static final String VALID_COURSE_BOB = "CS60";
-    public static final String VALID_GROUP_AMY = "T09";
-    public static final String VALID_GROUP_BOB = "T08";
-
-    public static final String ID_DESC_AMY = " " + PREFIX_CONTACT_ID_LONG + VALID_ID_AMY;
-    public static final String ID_DESC_BOB = " " + PREFIX_CONTACT_ID_LONG + VALID_ID_BOB;
-    public static final String NAME_DESC_AMY = " " + PREFIX_CONTACT_NAME_LONG + VALID_NAME_AMY;
-    public static final String NAME_DESC_BOB = " " + PREFIX_CONTACT_NAME_LONG + VALID_NAME_BOB;
-    public static final String EMAIL_DESC_AMY = " " + PREFIX_CONTACT_EMAIL_LONG + VALID_EMAIL_AMY;
-    public static final String EMAIL_DESC_BOB = " " + PREFIX_CONTACT_EMAIL_LONG + VALID_EMAIL_BOB;
-    public static final String TAG_DESC_FRIEND = " " + PREFIX_CONTACT_TAG_LONG + VALID_TAG_FRIEND;
-    public static final String TAG_DESC_HUSBAND = " " + PREFIX_CONTACT_TAG_LONG + VALID_TAG_HUSBAND;
-    public static final String TAG_DESC_FRIEND_HUSBAND =
-            " " + PREFIX_CONTACT_TAG_LONG + VALID_TAG_FRIEND + " " + VALID_TAG_HUSBAND;
-    public static final String COURSE_DESC_AMY = " " + PREFIX_CONTACT_COURSE_LONG + VALID_COURSE_AMY;
-    public static final String COURSE_DESC_BOB = " " + PREFIX_CONTACT_COURSE_LONG + VALID_COURSE_BOB;
-    public static final String GROUP_DESC_AMY = " " + PREFIX_CONTACT_GROUP_LONG + VALID_GROUP_AMY;
-    public static final String GROUP_DESC_BOB = " " + PREFIX_CONTACT_GROUP_LONG + VALID_GROUP_BOB;
-
-    // '&' not allowed in names
-    public static final String INVALID_NAME_DESC = " " + PREFIX_CONTACT_NAME_LONG + "-James&";
-    // missing '@' symbol
-    public static final String INVALID_EMAIL_DESC = " " + PREFIX_CONTACT_EMAIL_LONG + "bob!yahoo";
-    // '*' not allowed in tags
-    public static final String INVALID_TAG_DESC = " " + PREFIX_CONTACT_TAG_LONG + "-hubby*";
-    public static final String INVALID_TAG_HUSBAND_DESC = " " + PREFIX_CONTACT_TAG_LONG
-            + "husband -hubby*";
-
-    public static final String PREAMBLE_WHITESPACE = "\t  \r  \n";
-    public static final String PREAMBLE_NON_EMPTY = "NonEmptyPreamble";
-
-    public static final EditContactDescriptor DESC_AMY;
-    public static final EditContactDescriptor DESC_BOB;
-
-    static {
-        DESC_AMY = new EditContactDescriptorBuilder().withName(VALID_NAME_AMY)
-                                                    .withEmail(VALID_EMAIL_AMY)
-                                                    .withTags(VALID_TAG_FRIEND)
-                                                    .build();
-        DESC_BOB = new EditContactDescriptorBuilder().withName(VALID_NAME_BOB)
-                                                    .withEmail(VALID_EMAIL_BOB)
-                                                    .withTags(VALID_TAG_HUSBAND, VALID_TAG_FRIEND)
-                                                    .build();
-    }
-
-    // For Todo
-    public static final String VALID_NAME_HW = "Do homework";
-    public static final String VALID_NAME_EX = "Do exercise";
-    public static final String VALID_DATETIME_1 = "2021-01-01 23:59";
-    public static final String VALID_DATETIME_2 = "2021-02-01 00:00";
-    public static final String VALID_LOCATION_HW = "Canvas";
-    public static final String VALID_LOCATION_EX = "Prof office";
-    public static final String VALID_TAG_HW = "Important";
-    public static final String VALID_TAG_EX = "Very important";
-
-    public static final String NAME_DESC_HW = " " + PREFIX_TODO_NAME_LONG + VALID_NAME_HW;
-    public static final String NAME_DESC_EX = " " + PREFIX_TODO_NAME_LONG + VALID_NAME_EX;
-    public static final String DEADLINE_DESC_HW = " " + PREFIX_TODO_DEADLINE_LONG + VALID_DATETIME_1;
-    public static final String DEADLINE_DESC_EX = " " + PREFIX_TODO_DEADLINE_LONG + VALID_DATETIME_2;
-    public static final String LOCATION_DESC_HW = " " + PREFIX_TODO_LOCATION_LONG + VALID_LOCATION_HW;
-    public static final String LOCATION_DESC_EX = " " + PREFIX_TODO_LOCATION_LONG + VALID_LOCATION_EX;
-    public static final String TAG_DESC_HW = " " + PREFIX_TODO_TAG_LONG + VALID_TAG_HW;
-    public static final String TAG_DESC_EX = " " + PREFIX_TODO_TAG_LONG + VALID_TAG_EX;
-
-    // & not allowed
-    public static final String INVALID_NAME_TODO = " " + PREFIX_TODO_NAME_LONG + "Stuff&stuff";
-    public static final String INVALID_NOT_DATETIME = " " + PREFIX_TODO_DEADLINE_LONG + "aaaa";
-    public static final String INVALID_DATETIME_INCORRECT_FORMAT = " " + PREFIX_TODO_DEADLINE_LONG + "23:59 25-12-31";
-    // empty
-    public static final String INVALID_LOCATION_TODO = " " + PREFIX_TODO_LOCATION_LONG + "";
-    // * not allowed
-    public static final String INVALID_TAG_TODO = " " + PREFIX_TODO_TAG_LONG + "hubby*";
-
-
-
 
     /**
      * Executes the given {@code command}, confirms that <br>
@@ -140,7 +37,7 @@ public class CommandTestUtil {
      * that takes a string {@code expectedMessage}.
      */
     public static void assertCommandSuccess(Command command, Model actualModel, String expectedMessage,
-            Model expectedModel) {
+                                            Model expectedModel) {
         CommandResult expectedCommandResult = new CommandResult(expectedMessage);
         assertCommandSuccess(command, actualModel, expectedCommandResult, expectedModel);
     }
@@ -163,22 +60,6 @@ public class CommandTestUtil {
         assertEquals(expectedContactManager, actualModel.getContactManagerAndList().getItemManager());
         assertEquals(expectedFilteredList,
                 actualModel.getContactManagerAndList().getFilteredItemsList());
-    }
-    /**
-     * Updates {@code model}'s filtered list to show only the contact at the given {@code targetIndex} in the
-     * {@code model}'s address book.
-     */
-    public static void showContactAtIndex(Model model, Index targetIndex) {
-        assertTrue(targetIndex.getZeroBased()
-                < model.getContactManagerAndList().getFilteredItemsList().size());
-
-        Contact contact = model.getContactManagerAndList().getFilteredItemsList()
-                .get(targetIndex.getZeroBased());
-        final String[] splitName = contact.getName().value.split("\\s+");
-        model.getContactManagerAndList()
-                .updateFilteredItemsList(new NamePredicate(Operator.AND, List.of(splitName[0])));
-
-        assertEquals(1, model.getContactManagerAndList().getFilteredItemsList().size());
     }
 
 }
