+++ src/test/java/seedu/address/logic/commands/update/EditContactDescriptorTest.java
@@ -3,11 +3,11 @@ package seedu.address.logic.commands.update;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_TAG_HUSBAND;
 
 import org.junit.jupiter.api.Test;
