+++ src/test/java/seedu/address/logic/parser/contact/AddContactCommandParserTest.java
@@ -1,28 +1,28 @@
 package seedu.address.logic.parser.contact;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.COURSE_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.COURSE_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.GROUP_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.GROUP_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.ID_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.ID_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_EMAIL_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NAME_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_NON_EMPTY;
-import static seedu.address.logic.commands.CommandTestUtil.PREAMBLE_WHITESPACE;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND_HUSBAND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_GROUP_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_BOB;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.COURSE_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.COURSE_DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.EMAIL_DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.GROUP_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.GROUP_DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.ID_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.ID_DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_EMAIL_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_NAME_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_TAG_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.NAME_DESC_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.PREAMBLE_NON_EMPTY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.PREAMBLE_WHITESPACE;
+import static seedu.address.logic.commands.ContactCommandTestUtil.TAG_DESC_FRIEND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.TAG_DESC_FRIEND_HUSBAND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_EMAIL_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_GROUP_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_NAME_BOB;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_TAG_FRIEND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT_COURSE_LONG;
