+++ src/test/java/seedu/address/logic/parser/contact/EditContactCommandParserTest.java
@@ -1,18 +1,18 @@
 package seedu.address.logic.parser.contact;
 
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.commands.CommandTestUtil.EMAIL_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_EMAIL_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_NAME_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.INVALID_TAG_HUSBAND_DESC;
-import static seedu.address.logic.commands.CommandTestUtil.NAME_DESC_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.TAG_DESC_FRIEND_HUSBAND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_ADDRESS_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_EMAIL_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_NAME_AMY;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_FRIEND;
-import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.EMAIL_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_EMAIL_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_NAME_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.INVALID_TAG_HUSBAND_DESC;
+import static seedu.address.logic.commands.ContactCommandTestUtil.NAME_DESC_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.TAG_DESC_FRIEND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.TAG_DESC_FRIEND_HUSBAND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_ADDRESS_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_EMAIL_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_NAME_AMY;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_TAG_FRIEND;
+import static seedu.address.logic.commands.ContactCommandTestUtil.VALID_TAG_HUSBAND;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseSuccess;
 import static seedu.address.logic.parser.ParserUtil.MESSAGE_INVALID_INDEX;
